Fix XML encoding of program

+++ code/simple.c
@@ -4,7 +4,7 @@
 int main()
 {
     float a = 4.5;
-    int b = a * 3 / 2;
+    float b = a * 3 / 2;
     
     
 }
